Add a ChurchNumerals namespace
