+++ lib/church_numerals/zero.rb
@@ -1,0 +1,7 @@
+module ChurchNumerals
+  class Zero
+    def call(value, &action)
+      value
+    end
+  end
+end
