--- lib/zero.rb
@@ -1,5 +1,0 @@
-class Zero
-  def call(value, &action)
-    value
-  end
-end
